requirements.txt
Pillow
classy_vision
datasets
einops
git+https://github.com/openai/CLIP.git
h5py
numpy
numpy<2
open_clip_torch
optuna
packaging
pandas
pillow-heif
safetensors
scikit-learn
timm
torch
torch==1.13.1
torchvision
torchvision==0.14.1
tqdm
xformers
plotly
